Update Camera UI
